feat/SCRUM-35

+++ boot/api/src/main/java/com/clip/api/payment/feign/dto/PaymentConfirmDto.java
@@ -3,16 +3,17 @@ package com.clip.api.payment.feign.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
 public class PaymentConfirmDto {
     private final String paymentKey;
     private final UUID orderId;
-    private final int amount;
+    private final BigDecimal amount;
 
     @Builder
-    public PaymentConfirmDto(String paymentKey, UUID orderId, int amount) {
+    public PaymentConfirmDto(String paymentKey, UUID orderId, BigDecimal amount) {
         this.paymentKey = paymentKey;
         this.orderId = orderId;
         this.amount = amount;
